+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the JES-Explorer will be documented in this file.
 - Highlight the Selected Jobs and Job Files
 - Update the Job tree when a job is deleted or cancelled
 - Added '*' support for job id
+- Added app2app arguments: 'expand' - boolean that says to expand the job. In a list of jobs, this expands the first result. 'showDD' - string that will auto-open any dataset definition with this name, when expanding job.
 
 ## <2.0.0>